Networking

+++ src/app/Object/Object.hpp
@@ -28,8 +28,8 @@ class OBJECT
 {
 private:
     Sprite sprite;
-    int x;
-    int y;
+    float x;
+    float y;
     SDL_Rect box;
 
 
@@ -42,11 +42,11 @@ public:
     bool check_bounds();
     void setSprite(Sprite *new_sprite);
 
-    void setPos(int x, int y);
-    void setX(int x);
-    void setY(int y);
-    int getX();
-    int getY();
+    void setPos(float x, float y);
+    void setX(float x);
+    void setY(float y);
+    float getX();
+    float getY();
 
     void setObstacleLocations(std::vector<SDL_Rect>* obstaclesLocs);
     
@@ -80,7 +80,7 @@ public:
          * @return true  - moved object succesfully 
          * @return false - failed to move object 
          */
-    virtual bool move(int x, int y) = 0;
+    virtual bool move(float x, float y) = 0;
 
     /**
          * @brief place/teleport an object to a 2D location
@@ -91,7 +91,7 @@ public:
          * @return true  - placed player succesfully 
          * @return false - failed to place player 
          */
-    virtual bool place(int x, int y) = 0;
+    virtual bool place(float x, float y) = 0;
     ~OBJECT(); //destructor
 };
